+++ scripts/assets/_TestUtilEngine_Version.ps1
@@ -1,12 +1,14 @@
-#Version 1.00
+#Version 1.01
+#https://github.com/texhex/testutil/wiki/_fwLinkScript
 
-#Script requires PowerShell 4.0 or higher - http://blogs.msdn.com/b/powershell/archive/2009/02/06/requires-your-scripts.aspx
+
+#This script requires PowerShell 4.0 or higher 
 #require -version 4.0
 
-#Guard against common code errors - http://technet.microsoft.com/en-us/library/ff730970.aspx
+#Guard against common code errors
 Set-StrictMode -version 2.0
 
-#Terminate on errors - http://blogs.technet.com/b/heyscriptingguy/archive/2010/03/08/hey-scripting-guy-march-8-2010.aspx
+#Terminate script on errors 
 $ErrorActionPreference = 'Stop'