+++ scripts/tests/_Executed as native process.ps1
@@ -1,4 +1,6 @@
-#v1.02
+#v1.03
+#https://github.com/texhex/testutil/wiki/_fwLinkScript
+
 
 #This script requires PowerShell 4.0 or higher 
 #require -version 4.0
@@ -7,4 +9,4 @@
 Set-StrictMode -version 2.0
 
 #Terminate script on errors 
-$ErrorActionPreference = 'Stop'#Default is MAJOR because we expect that Test-TUWoW returns false$Return = @{Data = "OK"; Name="Executed with correct bitness"; Text= "TestUtil is executed with the correct bitness (64 bit on a 64 bit OS)"}if (Test-TUWoW) {       #We are affected by WoW64. Report this as Fail (MAJOR issue)   $Return.Data="Fail"   $Return.Text="TestUtil was not started with correct bitness! Please start it using TestUtilLauncher.exe."}$Return
+$ErrorActionPreference = 'Stop'#Default is OK because we expect that Test-TUWoW returns false$Return = @{Data = "OK"; Name="Executed with correct bitness"; Text= "TestUtil is executed with the correct bitness (64 bit on a 64 bit OS)"}if (Test-TUWoW) {       #We are affected by WoW64. Report this as Fail (MAJOR issue)   $Return.Data="Fail"   $Return.Text="TestUtil was not started with correct bitness! Please start it using TestUtilLauncher.exe."}$Return