+++ scripts/tests/_WMI Computer Status OK.ps1
@@ -1,0 +1,12 @@
+#v1.02
+#https://github.com/texhex/testutil/wiki/_fwLinkScript
+
+
+#This script requires PowerShell 4.0 or higher 
+#require -version 4.0
+
+#Guard against common code errors
+Set-StrictMode -version 2.0
+
+#Terminate script on errors 
+$ErrorActionPreference = 'Stop'$Return = @{Data = "OK"; Name="WMI Computer status is healty"; Text= "Windows Management Instrumentation (WMI) reports this system is healthy"}$wmi = Get-WMIObject Win32_ComputerSystem -Property "Status"[string]$status= $wmi.Status#$status="arg!"if ($status -ne "OK") {   $Return.Data="Fail"   $Return.Text="Windows Management Instrumentation (WMI) class Win32_ComputerSystem reported status [$status]"}$Return